- Added Integration Tests - Re-Added NanoHTTPD License - Privatized classes
